[#213]refactor: 분리했던 부분 되돌림...
