+++ client/src/components/productList/ProductItemContainer/index.tsx
@@ -11,7 +11,7 @@ import ProductItem from '~/components/product/ProductItem';
 import { useHistory } from '~/core/Router';
 import FetchContext from '~/lib/contexts/fetchContext';
 import FilterContext from '~/lib/contexts/filterContext';
-import { ProductData } from '~/components/productList/ProductList';
+import { ProductData } from '~/pages/ProductList';
 import { FINISH_FETCH, INIT_FETCH, START_FETCH } from '~/stores/fetchModule';
 
 import S from './index.style';
